Bug fix
Fixed a bug that shows stats wrong

+++ LoopLearn/Backend/Database/UserRepository.cs
@@ -40,7 +40,7 @@ namespace LoopLearn.Backend.Database
 
             using var cmd = new SQLiteCommand(query, conn);
             cmd.Parameters.AddWithValue("@u", username);
-            int count = (int)cmd.ExecuteScalar();
+            long count = (long)cmd.ExecuteScalar();
 
             return count > 0;
         }
